[AST] Apply new data and element segments.

+++ lib/aot/compiler.cpp
@@ -107,7 +107,7 @@ struct SSVM::AOT::Compiler::CompileContext {
       false;
 #endif
   std::vector<const AST::FunctionType *> FunctionTypes;
-  std::vector<unsigned int> Elements;
+  std::vector<uint64_t> Elements;
   std::vector<
       std::tuple<unsigned int, llvm::Function *, SSVM::AST::CodeSegment *>>
       Functions;
@@ -1520,7 +1520,8 @@ private:
 
     std::vector<std::pair<size_t, llvm::Function *>> Table;
     for (uint32_t I = 0; I < Context.Elements.size(); ++I) {
-      const auto FuncIdx = Context.Elements[I];
+      /// TODO: Handle nullptr case.
+      const auto FuncIdx = static_cast<uint32_t>(Context.Elements[I]);
       const auto FuncTypeIndex2 = std::get<0>(Context.Functions[FuncIdx]);
       const auto &FuncType2 = *Context.FunctionTypes[FuncTypeIndex2];
       if (FuncTypeIndex == FuncTypeIndex2 || FuncType == FuncType2) {
@@ -2301,11 +2302,15 @@ Expect<void> Compiler::compile(const AST::TableSection &TableSection,
     }
     const uint64_t Offset =
         llvm::cast<llvm::ConstantInt>(*Temp)->getZExtValue();
+    /// TODO: Instantiation behavior changed. Need to run exprs to calc ref
+    /// value.
+    /*
     const auto &FuncIdxes = Element->getFuncIdxes();
     if (Elements.size() < Offset + FuncIdxes.size()) {
       Elements.resize(Offset + FuncIdxes.size());
     }
     std::copy(FuncIdxes.begin(), FuncIdxes.end(), Elements.begin() + Offset);
+    */
   }
   return {};
 }
